Split codex gateway helpers
