+++ src/api/codexGateway.ts
@@ -42,10 +42,6 @@ import type {
   UiThread,
   UiReviewAction,
   UiReviewActionLevel,
-  UiReviewFile,
-  UiReviewFinding,
-  UiReviewHunk,
-  UiReviewLine,
   UiReviewResult,
   UiReviewScope,
   UiReviewSnapshot,
@@ -56,6 +52,9 @@ import type {
   UiThreadAutomationStatus,
 } from '../types/codex'
 import { normalizePathForUi } from '../pathUtils.js'
+import { normalizeReviewSnapshot, parseReviewText, readLatestReviewItem } from './codexGatewayReview'
+import { normalizeDirectoryApp, normalizeDirectoryMcpServer, normalizeDirectoryPluginApp, normalizeDirectoryPluginSkill, normalizeDirectoryPluginSummary } from './codexGatewayDirectory'
+export { attachThreadTerminal, closeThreadTerminal, getThreadTerminalQuickCommands, getThreadTerminalSnapshot, getThreadTerminalStatus, resizeThreadTerminal, sendThreadTerminalInput } from './codexGatewayTerminal'
 
 type CurrentModelConfig = {
   model: string
@@ -854,205 +853,6 @@ export async function getOlderThreadMessages(threadId: string, beforeTurnId: str
   }
 }
 
-function normalizeReviewLine(value: unknown): UiReviewLine | null {
-  const record = asRecord(value)
-  if (!record) return null
-
-  const key = readString(record.key)
-  const text = typeof record.text === 'string' ? record.text : ''
-  const kind = readString(record.kind)
-  if (!key || !kind) return null
-  if (kind !== 'meta' && kind !== 'hunk' && kind !== 'add' && kind !== 'remove' && kind !== 'context') {
-    return null
-  }
-
-  return {
-    key,
-    kind,
-    text,
-    oldLine: readNumber(record.oldLine),
-    newLine: readNumber(record.newLine),
-  }
-}
-
-function normalizeReviewHunk(value: unknown): UiReviewHunk | null {
-  const record = asRecord(value)
-  if (!record) return null
-
-  const id = readString(record.id)
-  const header = typeof record.header === 'string' ? record.header : ''
-  const patch = typeof record.patch === 'string' ? record.patch : ''
-  if (!id) return null
-
-  return {
-    id,
-    header,
-    patch,
-    addedLineCount: readNumber(record.addedLineCount) ?? 0,
-    removedLineCount: readNumber(record.removedLineCount) ?? 0,
-    oldStart: readNumber(record.oldStart),
-    oldLineCount: readNumber(record.oldLineCount) ?? 0,
-    newStart: readNumber(record.newStart),
-    newLineCount: readNumber(record.newLineCount) ?? 0,
-    lines: Array.isArray(record.lines)
-      ? record.lines
-        .map((entry) => normalizeReviewLine(entry))
-        .filter((entry): entry is UiReviewLine => entry !== null)
-      : [],
-  }
-}
-
-function normalizeReviewFile(value: unknown): UiReviewFile | null {
-  const record = asRecord(value)
-  if (!record) return null
-
-  const id = readString(record.id)
-  const path = readString(record.path)
-  const absolutePath = readString(record.absolutePath)
-  const operation = readString(record.operation)
-  if (!id || !path || !absolutePath || !operation) return null
-  if (operation !== 'add' && operation !== 'delete' && operation !== 'update' && operation !== 'rename') {
-    return null
-  }
-
-  return {
-    id,
-    path,
-    absolutePath,
-    previousPath: readString(record.previousPath),
-    previousAbsolutePath: readString(record.previousAbsolutePath),
-    operation,
-    addedLineCount: readNumber(record.addedLineCount) ?? 0,
-    removedLineCount: readNumber(record.removedLineCount) ?? 0,
-    diff: typeof record.diff === 'string' ? record.diff : '',
-    hunks: Array.isArray(record.hunks)
-      ? record.hunks
-        .map((entry) => normalizeReviewHunk(entry))
-        .filter((entry): entry is UiReviewHunk => entry !== null)
-      : [],
-  }
-}
-
-function normalizeReviewSnapshot(payload: unknown): UiReviewSnapshot {
-  const envelope = asRecord(payload)
-  const data = asRecord(envelope?.data)
-  const summaryRecord = asRecord(data?.summary)
-  const scope = readString(data?.scope) === 'baseBranch' ? 'baseBranch' : 'workspace'
-  const workspaceView = readString(data?.workspaceView) === 'staged' ? 'staged' : 'unstaged'
-
-  return {
-    cwd: readString(data?.cwd) ?? '',
-    gitRoot: readString(data?.gitRoot),
-    isGitRepo: readBoolean(data?.isGitRepo) ?? false,
-    scope,
-    workspaceView,
-    baseBranch: readString(data?.baseBranch),
-    baseBranchOptions: Array.isArray(data?.baseBranchOptions)
-      ? data.baseBranchOptions
-        .map((entry) => readString(entry))
-        .filter((entry): entry is string => typeof entry === 'string' && entry.length > 0)
-      : [],
-    headBranch: readString(data?.headBranch),
-    mergeBaseSha: readString(data?.mergeBaseSha),
-    generatedAtIso: readString(data?.generatedAtIso) ?? '',
-    summary: {
-      fileCount: readNumber(summaryRecord?.fileCount) ?? 0,
-      addedLineCount: readNumber(summaryRecord?.addedLineCount) ?? 0,
-      removedLineCount: readNumber(summaryRecord?.removedLineCount) ?? 0,
-    },
-    files: Array.isArray(data?.files)
-      ? data.files
-        .map((entry) => normalizeReviewFile(entry))
-        .filter((entry): entry is UiReviewFile => entry !== null)
-      : [],
-  }
-}
-
-function parseReviewLocation(value: string): {
-  absolutePath: string | null
-  startLine: number | null
-  endLine: number | null
-} {
-  const trimmed = value.trim()
-  if (!trimmed) {
-    return { absolutePath: null, startLine: null, endLine: null }
-  }
-
-  const match = trimmed.match(/^(.*?):(\d+)-(\d+)$/u)
-  if (!match) {
-    return { absolutePath: trimmed || null, startLine: null, endLine: null }
-  }
-
-  return {
-    absolutePath: match[1]?.trim() || null,
-    startLine: Number(match[2]),
-    endLine: Number(match[3]),
-  }
-}
-
-function parseReviewText(reviewText: string): UiReviewResult {
-  const normalized = reviewText.replace(/\r\n/g, '\n').trim()
-  if (!normalized) {
-    return { reviewText: '', summary: '', findings: [] }
-  }
-
-  const markerIndex = normalized.search(/\n(?:Full review comments|Review comment):\n/iu)
-  const summary = markerIndex >= 0 ? normalized.slice(0, markerIndex).trim() : normalized
-  const findingsSection = markerIndex >= 0 ? normalized.slice(markerIndex).trim() : ''
-  const findings: UiReviewFinding[] = []
-
-  if (findingsSection) {
-    const body = findingsSection
-      .replace(/^(?:Full review comments|Review comment):\n*/iu, '')
-      .trim()
-    const matches = body.matchAll(/^- (.+?) — (.+)\n?((?:  .*(?:\n|$))*)/gmu)
-    let index = 0
-    for (const match of matches) {
-      const title = match[1]?.trim() ?? ''
-      const location = parseReviewLocation(match[2] ?? '')
-      const block = (match[0] ?? '').trim()
-      const findingBody = (match[3] ?? '')
-        .split('\n')
-        .map((line) => line.replace(/^  /u, ''))
-        .join('\n')
-        .trim()
-
-      findings.push({
-        id: `finding:${index}`,
-        title: title || `Finding ${index + 1}`,
-        body: findingBody,
-        path: location.absolutePath ? location.absolutePath.split('/').filter(Boolean).slice(-1)[0] ?? location.absolutePath : null,
-        absolutePath: location.absolutePath,
-        startLine: location.startLine,
-        endLine: location.endLine,
-        rawText: block,
-      })
-      index += 1
-    }
-  }
-
-  return {
-    reviewText: normalized,
-    summary,
-    findings,
-  }
-}
-
-function readLatestReviewItem(payload: ThreadReadResponse, type: 'enteredReviewMode' | 'exitedReviewMode'): string | null {
-  const turns = Array.isArray(payload.thread.turns) ? payload.thread.turns : []
-  for (let turnIndex = turns.length - 1; turnIndex >= 0; turnIndex -= 1) {
-    const turn = turns[turnIndex]
-    const items = Array.isArray(turn?.items) ? turn.items : []
-    for (let itemIndex = items.length - 1; itemIndex >= 0; itemIndex -= 1) {
-      const item = items[itemIndex]
-      if (item?.type !== type) continue
-      const review = typeof item.review === 'string' ? item.review.trim() : ''
-      if (review) return review
-    }
-  }
-  return null
-}
-
 export async function getThreadReviewResult(threadId: string): Promise<{
   enteredReviewLabel: string | null
   result: UiReviewResult | null
@@ -1243,96 +1043,6 @@ export function subscribeCodexNotifications(onNotification: (value: RpcNotificat
 }
 
 export type { RpcNotification }
-
-function normalizeThreadTerminalSession(value: unknown): ThreadTerminalSession | null {
-  const record = asRecord(value)
-  if (!record) return null
-  const id = readString(record.id)
-  const threadId = readString(record.threadId)
-  const cwd = readString(record.cwd)
-  const shell = readString(record.shell)
-  if (!id || !threadId || !cwd || !shell) return null
-  return {
-    id,
-    threadId,
-    cwd,
-    shell,
-    buffer: typeof record.buffer === 'string' ? record.buffer : '',
-    truncated: readBoolean(record.truncated) ?? false,
-  }
-}
-
-async function fetchTerminalJson(path: string, init?: RequestInit): Promise<unknown> {
-  const response = await fetch(path, init)
-  const payload = await response.json().catch(() => null)
-  if (!response.ok) {
-    throw new Error(extractErrorMessage(payload, `Terminal request failed with HTTP ${response.status}`))
-  }
-  return payload
-}
-
-export async function attachThreadTerminal(input: ThreadTerminalAttachInput): Promise<ThreadTerminalSession> {
-  const payload = await fetchTerminalJson('/codex-api/thread-terminal/attach', {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify(input),
-  })
-  const session = normalizeThreadTerminalSession(asRecord(payload)?.session)
-  if (!session) throw new Error('Terminal attach response was malformed')
-  return session
-}
-
-export async function getThreadTerminalStatus(): Promise<{ available: boolean, reason: string | null }> {
-  const payload = await fetchTerminalJson('/codex-api/thread-terminal/status')
-  const record = asRecord(payload)
-  return {
-    available: readBoolean(record?.available) ?? false,
-    reason: readString(record?.reason) || null,
-  }
-}
-
-export async function getThreadTerminalQuickCommands(cwd: string): Promise<ThreadTerminalQuickCommand[]> {
-  const payload = await fetchTerminalJson(`/codex-api/thread-terminal/quick-commands?cwd=${encodeURIComponent(cwd)}`)
-  const payloadRecord = asRecord(payload)
-  const rows: unknown[] = Array.isArray(payloadRecord?.commands) ? payloadRecord.commands : []
-  return rows.flatMap((row: unknown) => {
-    const record = asRecord(row)
-    const label = readString(record?.label)
-    const value = readString(record?.value)
-    const source = readString(record?.source)
-    if (!label || !value || (source !== 'package' && source !== 'script' && source !== 'make')) return []
-    return [{ label, value, source }]
-  })
-}
-
-export async function sendThreadTerminalInput(sessionId: string, data: string): Promise<void> {
-  await fetchTerminalJson('/codex-api/thread-terminal/input', {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify({ sessionId, data }),
-  })
-}
-
-export async function resizeThreadTerminal(sessionId: string, cols: number, rows: number): Promise<void> {
-  await fetchTerminalJson('/codex-api/thread-terminal/resize', {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify({ sessionId, cols, rows }),
-  })
-}
-
-export async function closeThreadTerminal(sessionId: string): Promise<void> {
-  await fetchTerminalJson('/codex-api/thread-terminal/close', {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify({ sessionId }),
-  })
-}
-
-export async function getThreadTerminalSnapshot(threadId: string): Promise<ThreadTerminalSession | null> {
-  const payload = await fetchTerminalJson(`/codex-api/thread-terminal-snapshot?threadId=${encodeURIComponent(threadId)}`)
-  return normalizeThreadTerminalSession(asRecord(payload)?.session)
-}
 
 export async function replyToServerRequest(
   id: number,
@@ -1935,166 +1645,6 @@ export async function getCurrentModelConfig(): Promise<CurrentModelConfig> {
   const reasoningEffort = normalizeReasoningEffort(payload.config.model_reasoning_effort)
   const speedMode = normalizeSpeedMode(payload.config.service_tier)
   return { model, providerId, reasoningEffort, speedMode }
-}
-
-function normalizeDirectoryPluginApp(value: unknown): DirectoryPluginAppSummary | null {
-  const record = asRecord(value)
-  if (!record) return null
-  const id = readString(record.id)
-  const name = readString(record.name)
-  if (!id || !name) return null
-  return {
-    id,
-    name,
-    description: readString(record.description) ?? '',
-    installUrl: readString(record.installUrl ?? record.install_url) ?? '',
-    needsAuth: readBoolean(record.needsAuth ?? record.needs_auth) ?? false,
-  }
-}
-
-function normalizeDirectoryPluginSkill(value: unknown): DirectoryPluginSkillSummary | null {
-  const record = asRecord(value)
-  if (!record) return null
-  const name = readString(record.name)
-  const path = readString(record.path)
-  if (!name || !path) return null
-  const iface = asRecord(record.interface)
-  return {
-    name,
-    path,
-    description: readString(record.description) ?? '',
-    enabled: readBoolean(record.enabled) ?? true,
-    displayName: readString(iface?.displayName ?? iface?.display_name) ?? name,
-    shortDescription: readString(record.shortDescription ?? record.short_description) ?? '',
-  }
-}
-
-function normalizeDirectoryPluginSummary(
-  value: unknown,
-  marketplace: { name?: string; displayName?: string; path?: string | null } = {},
-): DirectoryPluginSummary | null {
-  const record = asRecord(value)
-  if (!record) return null
-  const id = readString(record.id)
-  const name = readString(record.name)
-  if (!id || !name) return null
-  const iface = asRecord(record.interface)
-  const source = asRecord(record.source)
-  const sourceType = readString(source?.type) ?? ''
-  const sourcePath = readString(source?.path)
-  const sourceUrl = readString(source?.url) ?? ''
-  const remoteMarketplaceName = sourceType === 'remote' ? marketplace.name ?? null : null
-  const marketplacePath = marketplace.path ?? (sourceType === 'local' ? sourcePath : null)
-  const displayName = readString(iface?.displayName ?? iface?.display_name) ?? name
-  const shortDescription = readString(iface?.shortDescription ?? iface?.short_description)
-  const longDescription = readString(iface?.longDescription ?? iface?.long_description) ?? ''
-
-  return {
-    id,
-    name,
-    displayName,
-    description: shortDescription ?? longDescription,
-    longDescription,
-    developerName: readString(iface?.developerName ?? iface?.developer_name) ?? '',
-    category: readString(iface?.category) ?? '',
-    marketplaceName: marketplace.name ?? '',
-    marketplaceDisplayName: marketplace.displayName ?? marketplace.name ?? '',
-    marketplacePath,
-    remoteMarketplaceName,
-    sourceType,
-    sourceUrl,
-    installed: readBoolean(record.installed) ?? false,
-    enabled: readBoolean(record.enabled) ?? true,
-    installPolicy: readString(record.installPolicy ?? record.install_policy) ?? '',
-    authPolicy: readString(record.authPolicy ?? record.auth_policy) ?? '',
-    logoUrl: readString(iface?.logoUrl ?? iface?.logo_url) ?? '',
-    logoPath: readString(iface?.logo) ?? '',
-    composerIconUrl: readString(iface?.composerIconUrl ?? iface?.composer_icon_url) ?? '',
-    composerIconPath: readString(iface?.composerIcon ?? iface?.composer_icon) ?? '',
-    brandColor: readString(iface?.brandColor ?? iface?.brand_color) ?? '',
-    capabilities: readStringArray(iface?.capabilities),
-    defaultPrompt: readStringArray(iface?.defaultPrompt ?? iface?.default_prompt),
-    screenshotUrls: readStringArray(iface?.screenshotUrls ?? iface?.screenshot_urls),
-    screenshots: readStringArray(iface?.screenshots),
-    websiteUrl: readString(iface?.websiteUrl ?? iface?.website_url) ?? '',
-    privacyPolicyUrl: readString(iface?.privacyPolicyUrl ?? iface?.privacy_policy_url) ?? '',
-    termsOfServiceUrl: readString(iface?.termsOfServiceUrl ?? iface?.terms_of_service_url) ?? '',
-  }
-}
-
-function normalizeDirectoryApp(value: unknown, catalogRank = 0): DirectoryAppInfo | null {
-  const record = asRecord(value)
-  if (!record) return null
-  const id = readString(record.id)
-  const name = readString(record.name)
-  if (!id || !name) return null
-  const branding = asRecord(record.branding)
-  const metadata = asRecord(record.appMetadata ?? record.app_metadata)
-  return {
-    id,
-    name,
-    description: readString(record.description) ?? readString(metadata?.seoDescription ?? metadata?.seo_description) ?? '',
-    logoUrl: readString(record.logoUrl ?? record.logo_url) ?? '',
-    logoUrlDark: readString(record.logoUrlDark ?? record.logo_url_dark) ?? '',
-    distributionChannel: readString(record.distributionChannel ?? record.distribution_channel) ?? '',
-    installUrl: readString(record.installUrl ?? record.install_url) ?? '',
-    isAccessible: readBoolean(record.isAccessible ?? record.is_accessible) ?? false,
-    isEnabled: readBoolean(record.isEnabled ?? record.is_enabled) ?? true,
-    pluginDisplayNames: readStringArray(record.pluginDisplayNames ?? record.plugin_display_names),
-    category: readString(branding?.category) ?? '',
-    developer: readString(branding?.developer) ?? readString(metadata?.developer) ?? '',
-    website: readString(branding?.website) ?? '',
-    privacyPolicy: readString(branding?.privacyPolicy ?? branding?.privacy_policy) ?? '',
-    termsOfService: readString(branding?.termsOfService ?? branding?.terms_of_service) ?? '',
-    catalogRank,
-  }
-}
-
-function normalizeDirectoryMcpServer(value: unknown): DirectoryMcpServerStatus | null {
-  const record = asRecord(value)
-  if (!record) return null
-  const name = readString(record.name)
-  if (!name) return null
-  const toolsRecord = asRecord(record.tools) ?? {}
-  const tools = Object.entries(toolsRecord).map(([fallbackName, raw]) => {
-    const tool = asRecord(raw)
-    return {
-      name: readString(tool?.name) ?? fallbackName,
-      title: readString(tool?.title) ?? '',
-      description: readString(tool?.description) ?? '',
-    }
-  })
-  const resources = Array.isArray(record.resources)
-    ? record.resources.map((raw) => {
-      const resource = asRecord(raw)
-      return {
-        name: readString(resource?.name) ?? '',
-        title: readString(resource?.title) ?? '',
-        uri: readString(resource?.uri) ?? '',
-        description: readString(resource?.description) ?? '',
-      }
-    }).filter((resource) => resource.name || resource.uri)
-    : []
-  const rawResourceTemplates = record.resourceTemplates ?? record.resource_templates
-  const resourceTemplates = Array.isArray(rawResourceTemplates)
-    ? rawResourceTemplates.map((raw: unknown) => {
-      const template = asRecord(raw)
-      return {
-        name: readString(template?.name) ?? '',
-        title: readString(template?.title) ?? '',
-        uriTemplate: readString(template?.uriTemplate ?? template?.uri_template) ?? '',
-        description: readString(template?.description) ?? '',
-      }
-    }).filter((template) => template.name || template.uriTemplate)
-    : []
-
-  return {
-    name,
-    authStatus: readString(record.authStatus ?? record.auth_status) ?? 'unsupported',
-    tools,
-    resources,
-    resourceTemplates,
-  }
 }
 
 export async function listDirectoryPlugins(cwds?: string[]): Promise<DirectoryPluginSummary[]> {
