+++ src/api/codexGatewayReview.ts
@@ -1,0 +1,226 @@
+import type { ThreadReadResponse } from './appServerDtos'
+import type {
+  UiReviewFile,
+  UiReviewFinding,
+  UiReviewHunk,
+  UiReviewLine,
+  UiReviewResult,
+  UiReviewSnapshot,
+} from '../types/codex'
+
+function asRecord(value: unknown): Record<string, unknown> | null {
+  return value !== null && typeof value === 'object' && !Array.isArray(value)
+    ? (value as Record<string, unknown>)
+    : null
+}
+
+function readString(value: unknown): string | null {
+  return typeof value === 'string' && value.length > 0 ? value : null
+}
+
+function readNumber(value: unknown): number | null {
+  return typeof value === 'number' && Number.isFinite(value) ? value : null
+}
+
+function readBoolean(value: unknown): boolean | null {
+  return typeof value === 'boolean' ? value : null
+}
+
+function normalizeReviewLine(value: unknown): UiReviewLine | null {
+  const record = asRecord(value)
+  if (!record) return null
+
+  const key = readString(record.key)
+  const text = typeof record.text === 'string' ? record.text : ''
+  const kind = readString(record.kind)
+  if (!key || !kind) return null
+  if (kind !== 'meta' && kind !== 'hunk' && kind !== 'add' && kind !== 'remove' && kind !== 'context') {
+    return null
+  }
+
+  return {
+    key,
+    kind,
+    text,
+    oldLine: readNumber(record.oldLine),
+    newLine: readNumber(record.newLine),
+  }
+}
+
+function normalizeReviewHunk(value: unknown): UiReviewHunk | null {
+  const record = asRecord(value)
+  if (!record) return null
+
+  const id = readString(record.id)
+  const header = typeof record.header === 'string' ? record.header : ''
+  const patch = typeof record.patch === 'string' ? record.patch : ''
+  if (!id) return null
+
+  return {
+    id,
+    header,
+    patch,
+    addedLineCount: readNumber(record.addedLineCount) ?? 0,
+    removedLineCount: readNumber(record.removedLineCount) ?? 0,
+    oldStart: readNumber(record.oldStart),
+    oldLineCount: readNumber(record.oldLineCount) ?? 0,
+    newStart: readNumber(record.newStart),
+    newLineCount: readNumber(record.newLineCount) ?? 0,
+    lines: Array.isArray(record.lines)
+      ? record.lines
+        .map((entry) => normalizeReviewLine(entry))
+        .filter((entry): entry is UiReviewLine => entry !== null)
+      : [],
+  }
+}
+
+function normalizeReviewFile(value: unknown): UiReviewFile | null {
+  const record = asRecord(value)
+  if (!record) return null
+
+  const id = readString(record.id)
+  const path = readString(record.path)
+  const absolutePath = readString(record.absolutePath)
+  const operation = readString(record.operation)
+  if (!id || !path || !absolutePath || !operation) return null
+  if (operation !== 'add' && operation !== 'delete' && operation !== 'update' && operation !== 'rename') {
+    return null
+  }
+
+  return {
+    id,
+    path,
+    absolutePath,
+    previousPath: readString(record.previousPath),
+    previousAbsolutePath: readString(record.previousAbsolutePath),
+    operation,
+    addedLineCount: readNumber(record.addedLineCount) ?? 0,
+    removedLineCount: readNumber(record.removedLineCount) ?? 0,
+    diff: typeof record.diff === 'string' ? record.diff : '',
+    hunks: Array.isArray(record.hunks)
+      ? record.hunks
+        .map((entry) => normalizeReviewHunk(entry))
+        .filter((entry): entry is UiReviewHunk => entry !== null)
+      : [],
+  }
+}
+
+export function normalizeReviewSnapshot(payload: unknown): UiReviewSnapshot {
+  const envelope = asRecord(payload)
+  const data = asRecord(envelope?.data)
+  const summaryRecord = asRecord(data?.summary)
+  const scope = readString(data?.scope) === 'baseBranch' ? 'baseBranch' : 'workspace'
+  const workspaceView = readString(data?.workspaceView) === 'staged' ? 'staged' : 'unstaged'
+
+  return {
+    cwd: readString(data?.cwd) ?? '',
+    gitRoot: readString(data?.gitRoot),
+    isGitRepo: readBoolean(data?.isGitRepo) ?? false,
+    scope,
+    workspaceView,
+    baseBranch: readString(data?.baseBranch),
+    baseBranchOptions: Array.isArray(data?.baseBranchOptions)
+      ? data.baseBranchOptions
+        .map((entry) => readString(entry))
+        .filter((entry): entry is string => typeof entry === 'string' && entry.length > 0)
+      : [],
+    headBranch: readString(data?.headBranch),
+    mergeBaseSha: readString(data?.mergeBaseSha),
+    generatedAtIso: readString(data?.generatedAtIso) ?? '',
+    summary: {
+      fileCount: readNumber(summaryRecord?.fileCount) ?? 0,
+      addedLineCount: readNumber(summaryRecord?.addedLineCount) ?? 0,
+      removedLineCount: readNumber(summaryRecord?.removedLineCount) ?? 0,
+    },
+    files: Array.isArray(data?.files)
+      ? data.files
+        .map((entry) => normalizeReviewFile(entry))
+        .filter((entry): entry is UiReviewFile => entry !== null)
+      : [],
+  }
+}
+
+function parseReviewLocation(value: string): {
+  absolutePath: string | null
+  startLine: number | null
+  endLine: number | null
+} {
+  const trimmed = value.trim()
+  if (!trimmed) {
+    return { absolutePath: null, startLine: null, endLine: null }
+  }
+
+  const match = trimmed.match(/^(.*?):(\d+)-(\d+)$/u)
+  if (!match) {
+    return { absolutePath: trimmed || null, startLine: null, endLine: null }
+  }
+
+  return {
+    absolutePath: match[1]?.trim() || null,
+    startLine: Number(match[2]),
+    endLine: Number(match[3]),
+  }
+}
+
+export function parseReviewText(reviewText: string): UiReviewResult {
+  const normalized = reviewText.replace(/\r\n/g, '\n').trim()
+  if (!normalized) {
+    return { reviewText: '', summary: '', findings: [] }
+  }
+
+  const markerIndex = normalized.search(/\n(?:Full review comments|Review comment):\n/iu)
+  const summary = markerIndex >= 0 ? normalized.slice(0, markerIndex).trim() : normalized
+  const findingsSection = markerIndex >= 0 ? normalized.slice(markerIndex).trim() : ''
+  const findings: UiReviewFinding[] = []
+
+  if (findingsSection) {
+    const body = findingsSection
+      .replace(/^(?:Full review comments|Review comment):\n*/iu, '')
+      .trim()
+    const matches = body.matchAll(/^- (.+?) — (.+)\n?((?:  .*(?:\n|$))*)/gmu)
+    let index = 0
+    for (const match of matches) {
+      const title = match[1]?.trim() ?? ''
+      const location = parseReviewLocation(match[2] ?? '')
+      const block = (match[0] ?? '').trim()
+      const findingBody = (match[3] ?? '')
+        .split('\n')
+        .map((line) => line.replace(/^  /u, ''))
+        .join('\n')
+        .trim()
+
+      findings.push({
+        id: `finding:${index}`,
+        title: title || `Finding ${index + 1}`,
+        body: findingBody,
+        path: location.absolutePath ? location.absolutePath.split('/').filter(Boolean).slice(-1)[0] ?? location.absolutePath : null,
+        absolutePath: location.absolutePath,
+        startLine: location.startLine,
+        endLine: location.endLine,
+        rawText: block,
+      })
+      index += 1
+    }
+  }
+
+  return {
+    reviewText: normalized,
+    summary,
+    findings,
+  }
+}
+
+export function readLatestReviewItem(payload: ThreadReadResponse, type: 'enteredReviewMode' | 'exitedReviewMode'): string | null {
+  const turns = Array.isArray(payload.thread.turns) ? payload.thread.turns : []
+  for (let turnIndex = turns.length - 1; turnIndex >= 0; turnIndex -= 1) {
+    const turn = turns[turnIndex]
+    const items = Array.isArray(turn?.items) ? turn.items : []
+    for (let itemIndex = items.length - 1; itemIndex >= 0; itemIndex -= 1) {
+      const item = items[itemIndex]
+      if (item?.type !== type) continue
+      const review = typeof item.review === 'string' ? item.review.trim() : ''
+      if (review) return review
+    }
+  }
+  return null
+}
